Rename to metrics-exporter, add Dockerfile and better cli parsing
